requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.1
django-admin-material-dashboard==1.0.10
django-ajax==0.3.0
django-allauth==0.61.1
django-appconf==1.0.6
django-ckeditor-5==0.2.11
django-countries==7.5.1
django-crispy-forms==2.1
django-datatable-view==2.1.6
django-datatables-view==1.20.0
django-extensions==3.2.3
django-htmx==1.17.2
django-phonenumber-field==7.3.0
django-select2==8.1.2
djangorestframework==3.14.0
Faker==22.7.0
gunicorn==21.2.0
idna==3.6
install==1.3.5
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.29
Pillow==9.5.0
psycopg2-binary==2.9.9
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
stripe==8.2.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
